WIP: http: use less space for ValidUntil

+++ aquatic_http/src/config.rs
@@ -112,9 +112,9 @@ pub struct CleaningConfig {
     /// Clean connections this often (seconds)
     pub connection_cleaning_interval: u64,
     /// Remove peers that have not announced for this long (seconds)
-    pub max_peer_age: u64,
+    pub max_peer_age: u32,
     /// Remove connections that haven't seen valid requests for this long (seconds)
-    pub max_connection_idle: u64,
+    pub max_connection_idle: u32,
 }
 
 impl Default for CleaningConfig {
